refactor(admin): extract Settings > Reports panel into standalone component
Move the Settings > Reports (weekly/monthly scheduled email) panel out of
the monolithic AdminComponent into a standalone OnPush AdminReportsComponent
(selector app-admin-reports). The component owns the reports schedule form +
last-sent/last-error status and the load/save (conflict-aware, 404 create
fallback) and send-now behaviour; the shared CMS content-version bookkeeping
stays on the parent and is threaded in via the rememberContentVersion /
withExpectedVersion / handleContentConflict / forgetContentVersion callbacks.
TranslateService stays threaded. Behavioural specs move with the code from the
parent spec suites into admin-reports.component.spec.ts.
