+++ frontend/src/app/pages/admin/admin.component.branchfill.spec.ts
@@ -1,4 +1,3 @@
- 
 import { DomSanitizer } from '@angular/platform-browser';
 import { ActivatedRoute } from '@angular/router';
 import { Subject, of, throwError } from 'rxjs';
@@ -2781,137 +2780,6 @@ describe('AdminComponent — branch fill', () => {
       c.updateCategoryTaxGroup(fail, 'new');
       expect(fail.tax_group_id).toBe('old');
       expect(h.toast.error).toHaveBeenCalledWith('adminUi.taxes.errors.categoryAssign');
-    });
-  });
-
-  describe('reports settings', () => {
-    it('loadReportsSettings parses string flags, list/recipients and last-run markers', () => {
-      h.admin.getContent.and.returnValue(
-        of({
-          meta: {
-            reports_weekly_enabled: 'yes',
-            reports_monthly_enabled: 'off',
-            reports_weekly_weekday: '3',
-            reports_weekly_hour_utc: '30',
-            reports_monthly_day: '50',
-            reports_recipients: ['a@b.com', '  ', 'c@d.com'],
-            reports_weekly_last_sent_period_end: '2026-01-01',
-            reports_weekly_last_error: 'werr',
-            reports_monthly_last_sent_period_end: '2026-02-01',
-            reports_monthly_last_error: 'merr',
-          },
-        }),
-      );
-      c.loadReportsSettings();
-      expect(c.reportsSettingsForm.weekly_enabled).toBe(true);
-      expect(c.reportsSettingsForm.monthly_enabled).toBe(false);
-      expect(c.reportsSettingsForm.weekly_weekday).toBe(3);
-      expect(c.reportsSettingsForm.weekly_hour_utc).toBe(23);
-      expect(c.reportsSettingsForm.monthly_day).toBe('28');
-      expect(c.reportsSettingsForm.recipients).toBe('a@b.com, c@d.com');
-      expect(c.reportsMonthlyLastSent).toBe('2026-02-01');
-      expect(c.reportsMonthlyLastError).toBe('merr');
-    });
-
-    it('loadReportsSettings parses a string recipient list and resets on error', () => {
-      h.admin.getContent.and.returnValue(
-        of({ meta: { reports_recipients: 'a@b.com; c@d.com\ne@f.com' } }),
-      );
-      c.loadReportsSettings();
-      expect(c.reportsSettingsForm.recipients).toBe('a@b.com, c@d.com, e@f.com');
-
-      c.contentVersions['site.reports'] = { version: 1 };
-      h.admin.getContent.and.returnValue(throwError(() => ({})));
-      c.loadReportsSettings();
-      expect(c.contentVersions['site.reports']).toBeUndefined();
-      expect(c.reportsSettingsForm.recipients).toBe('');
-    });
-
-    it('saveReportsSettings normalises numbers, dedupes valid recipients and persists', () => {
-      c.reportsSettingsMeta = {};
-      c.reportsSettingsForm = {
-        weekly_enabled: true,
-        weekly_weekday: 0,
-        weekly_hour_utc: 0,
-        monthly_enabled: false,
-        monthly_day: '',
-        monthly_hour_utc: 0,
-        recipients: 'a@b.com, a@b.com, bad, c@d.com',
-      };
-      h.admin.updateContentBlock.and.returnValue(
-        of({ meta: { reports_recipients: ['a@b.com', 'c@d.com'] } }),
-      );
-      c.saveReportsSettings();
-      const meta = h.admin.updateContentBlock.calls.mostRecent().args[1].meta;
-      expect(meta.reports_monthly_day).toBe(1);
-      expect(meta.reports_recipients).toEqual(['a@b.com', 'c@d.com']);
-      expect(meta.reports_top_products_limit).toBe(5);
-      expect(c.reportsSettingsMessage).toBe('adminUi.reports.success.save');
-    });
-
-    it('saveReportsSettings drops recipients when none are valid and keeps preset limits', () => {
-      c.reportsSettingsMeta = {
-        reports_top_products_limit: 9,
-        reports_low_stock_limit: 9,
-        reports_retry_cooldown_minutes: 9,
-      };
-      c.reportsSettingsForm = {
-        weekly_enabled: false,
-        weekly_weekday: 9,
-        weekly_hour_utc: 9,
-        monthly_enabled: true,
-        monthly_day: '15',
-        monthly_hour_utc: 9,
-        recipients: 'not-an-email',
-      };
-      h.admin.updateContentBlock.and.returnValue(of(null));
-      c.saveReportsSettings();
-      const meta = h.admin.updateContentBlock.calls.mostRecent().args[1].meta;
-      expect('reports_recipients' in meta).toBe(false);
-      expect(meta.reports_top_products_limit).toBe(9);
-    });
-
-    it('saveReportsSettings surfaces a conflict, then creates on a generic error', () => {
-      c.reportsSettingsMeta = {};
-      c.reportsSettingsForm = {
-        weekly_enabled: false,
-        weekly_weekday: 0,
-        weekly_hour_utc: 8,
-        monthly_enabled: false,
-        monthly_day: '1',
-        monthly_hour_utc: 8,
-        recipients: '',
-      };
-      h.admin.updateContentBlock.and.returnValue(throwError(() => ({ status: 409 })));
-      h.admin.getContent.and.returnValue(of({ meta: {} }));
-      c.saveReportsSettings();
-      expect(c.reportsSettingsError).toBe('adminUi.reports.errors.save');
-
-      h.admin.updateContentBlock.and.returnValue(throwError(() => ({ status: 500 })));
-      h.admin.createContent.and.returnValue(of({ meta: {} }));
-      c.saveReportsSettings();
-      expect(c.reportsSettingsMessage).toBe('adminUi.reports.success.save');
-
-      h.admin.createContent.and.returnValue(throwError(() => ({})));
-      c.saveReportsSettings();
-      expect(c.reportsSettingsError).toBe('adminUi.reports.errors.save');
-    });
-
-    it('sendReportNow guards re-entry and maps skipped/sent/error', () => {
-      c.reportsSending = true;
-      c.sendReportNow('weekly');
-      expect(h.admin.sendScheduledReport).not.toHaveBeenCalled();
-      c.reportsSending = false;
-      spyOn(c, 'loadReportsSettings');
-      h.admin.sendScheduledReport.and.returnValue(of({ skipped: true }));
-      c.sendReportNow('weekly');
-      expect(c.reportsSettingsMessage).toBe('adminUi.reports.success.skipped');
-      h.admin.sendScheduledReport.and.returnValue(of({ skipped: false }));
-      c.sendReportNow('monthly', true);
-      expect(c.reportsSettingsMessage).toBe('adminUi.reports.success.sent');
-      h.admin.sendScheduledReport.and.returnValue(throwError(() => ({})));
-      c.sendReportNow('weekly');
-      expect(c.reportsSettingsError).toBe('adminUi.reports.errors.send');
     });
   });
