+++ frontend/src/app/pages/admin/admin.component.coverage.spec.ts
@@ -2589,109 +2589,6 @@ describe('AdminComponent — blog image markdown, a11y and writing aids', () => 
   });
 });
 
-describe('AdminComponent — settings save/load (checkout, reports, assets)', () => {
-  let h: Harness;
-  let c: any;
-  beforeEach(() => {
-    h = createComponent();
-    c = h.component as any;
-    h.admin.getContent.and.returnValue(of({ meta: {}, version: 1 })); // safe reload
-  });
-
-  it('loadReportsSettings parses meta and resets on error', () => {
-    h.admin.getContent.and.returnValue(
-      of({
-        version: 3,
-        meta: {
-          reports_weekly_enabled: 'yes',
-          reports_weekly_weekday: 9,
-          reports_weekly_hour_utc: 30,
-          reports_monthly_enabled: 1,
-          reports_monthly_day: 40,
-          reports_monthly_hour_utc: '5',
-          reports_recipients: 'a@b.com, bad, c@d.com',
-          reports_weekly_last_sent_period_end: '2030',
-          reports_weekly_last_error: 'err',
-        },
-      }),
-    );
-    c.loadReportsSettings();
-    expect(c.reportsSettingsForm.weekly_enabled).toBe(true);
-    expect(c.reportsSettingsForm.weekly_weekday).toBe(6); // clamped
-    expect(c.reportsSettingsForm.weekly_hour_utc).toBe(23); // clamped
-    expect(c.reportsSettingsForm.monthly_day).toBe('28'); // clamped
-    expect(c.reportsSettingsForm.recipients).toContain('a@b.com');
-    expect(c.reportsWeeklyLastSent).toBe('2030');
-
-    h.admin.getContent.and.returnValue(of({ meta: { reports_recipients: ['x@y.com', ''] } }));
-    c.loadReportsSettings();
-    expect(c.reportsSettingsForm.recipients).toBe('x@y.com');
-
-    h.admin.getContent.and.returnValue(throwError(() => new Error('x')));
-    c.loadReportsSettings();
-    expect(c.reportsSettingsForm.recipients).toBe('');
-  });
-
-  it('saveReportsSettings filters recipients and persists', () => {
-    c.reportsSettingsMeta = {};
-    c.reportsSettingsForm = {
-      weekly_enabled: true,
-      weekly_weekday: 2,
-      weekly_hour_utc: 9,
-      monthly_enabled: false,
-      monthly_day: '15',
-      monthly_hour_utc: 8,
-      recipients: 'a@b.com, a@b.com, invalid, c@d.com',
-    };
-    h.admin.updateContentBlock.and.returnValue(
-      of({ version: 2, meta: { reports_recipients: ['a@b.com', 'c@d.com'] } }),
-    );
-    c.saveReportsSettings();
-    const payload = h.admin.updateContentBlock.calls.mostRecent().args[1];
-    expect(payload.meta.reports_recipients).toEqual(['a@b.com', 'c@d.com']);
-    expect(payload.meta.reports_top_products_limit).toBe(5);
-    expect(c.reportsSettingsMessage).toBeTruthy();
-
-    // no valid recipients deletes the key
-    c.reportsSettingsForm.recipients = 'none';
-    c.saveReportsSettings();
-    const p2 = h.admin.updateContentBlock.calls.mostRecent().args[1];
-    expect('reports_recipients' in p2.meta).toBe(false);
-
-    h.admin.updateContentBlock.and.returnValue(throwError(() => ({ status: 409 })));
-    c.saveReportsSettings();
-    expect(c.reportsSettingsError).toBeTruthy();
-
-    h.admin.updateContentBlock.and.returnValue(throwError(() => ({ status: 500 })));
-    h.admin.createContent.and.returnValue(of({ version: 1, meta: {} }));
-    c.saveReportsSettings();
-    expect(c.reportsSettingsMessage).toBeTruthy();
-    h.admin.createContent.and.returnValue(throwError(() => new Error('x')));
-    c.saveReportsSettings();
-    expect(c.reportsSettingsError).toBeTruthy();
-  });
-
-  it('sendReportNow guards re-entry and reports skip/sent/error', () => {
-    spyOn(c, 'loadReportsSettings'); // reload would clear the success message
-    c.reportsSending = true;
-    c.sendReportNow('weekly');
-    expect(h.admin.sendScheduledReport).not.toHaveBeenCalled();
-
-    c.reportsSending = false;
-    h.admin.sendScheduledReport.and.returnValue(of({ skipped: true }));
-    c.sendReportNow('weekly');
-    expect(c.reportsSettingsMessage).toContain('skipped');
-
-    h.admin.sendScheduledReport.and.returnValue(of({ skipped: false }));
-    c.sendReportNow('monthly', true);
-    expect(c.reportsSettingsMessage).toContain('sent');
-
-    h.admin.sendScheduledReport.and.returnValue(throwError(() => new Error('x')));
-    c.sendReportNow('weekly');
-    expect(c.reportsSettingsError).toBeTruthy();
-  });
-});
-
 describe('AdminComponent — site settings load/save (assets, social, seo)', () => {
   let h: Harness;
   let c: any;
